Update event and rename classes
