+++ src/main/java/com/rackspace/salus/telemetry/entities/StateChange.java
@@ -1,0 +1,76 @@
+/*
+ * Copyright 2020 Rackspace US, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.rackspace.salus.telemetry.entities;
+
+import java.io.Serializable;
+import java.time.Instant;
+import java.util.UUID;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+import javax.persistence.Table;
+import lombok.Data;
+import org.hibernate.annotations.CreationTimestamp;
+import org.hibernate.annotations.UpdateTimestamp;
+
+@Entity
+@Table
+@Data
+public class StateChange implements Serializable {
+
+  @Id
+  @GeneratedValue
+  @org.hibernate.annotations.Type(type="uuid-char")
+  UUID id;
+
+  @Column(name="tenant_id")
+  String tenantId;
+
+  @Column(name="resource_id")
+  String resourceId;
+
+  @Column(name="monitor_id")
+  UUID monitorId;
+
+  @Column(name="task_id")
+  UUID taskId;
+
+  @Column(name="state")
+  String state;
+
+  @Column(name="message")
+  String message;
+
+  // a serialized list of EnrichedMetric
+  @Column(name="contributing_events")
+  String contributingEvents;
+
+  @Column(name="previous_state")
+  String previousState;
+
+  @Column(name="evaluation_timestamp")
+  Instant evaluationTimestamp;
+
+  @CreationTimestamp
+  @Column(name="created_timestamp")
+  Instant createdTimestamp;
+
+  @UpdateTimestamp
+  @Column(name="updated_timestamp")
+  Instant updatedTimestamp;
+}
